borken kernel test fix plus comments application
